requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13.0
psycopg2-binary>=2.9.3
django-cors-headers>=3.10.0